requirements/base.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
attrs==22.1.0
black==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.4
Django==4.1.4
django-cors-headers==3.13.0
django-environ==0.9.0
django-phonenumber-field==7.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==22.0
pathspec==0.10.2
phonenumberslite==8.13.1
Pillow==9.3.0
platformdirs==2.6.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyrsistent==0.19.2
pyshorteners==1.0.1
pytz==2022.6
PyYAML==6.0
qrcode==7.3.1
requests==2.28.1
ruamel.yaml==0.17.21
sqlparse==0.4.3
swagger-spec-validator==3.0.3
typing_extensions==4.4.0
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.13
